pyproject.toml
[project]
name = "copilot-evals"
version = "0.1.0"
description = "A/B evaluation framework for GitHub Copilot customizations using OTel telemetry"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31",
    "click>=8.1",
]

[project.scripts]
copilot-eval = "eval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eval"]
